wallet: split itest GetAccount tests
